fix some bug, add ERC20

+++ contracts/NFT.sol
@@ -1,7 +1,7 @@
 //SPDX-License-Identifier: MIT
 pragma solidity ^0.8.4;
 
-// we will bring in the openzeppelin ERC721 NFT functionality 
+// we will bring in the openzeppelin ERC721 NFT functionality
 
 import '@openzeppelin/contracts/token/ERC721/ERC721.sol';
 import '@openzeppelin/contracts/token/ERC721/extensions/ERC721URIStorage.sol';
@@ -16,14 +16,15 @@ contract NFT is ERC721URIStorage {
     address contractAddress;
 
     // OBJ: give the NFT market the ability to transact with tokens or change ownership
-    // setApprovalForAll allows us to do that with contract address 
+    // setApprovalForAll allows us to do that with contract address
 
     // constructor set up our address
     constructor(address marketplaceAddress) ERC721('CryptoLion', 'CLION') {
         contractAddress = marketplaceAddress;
     }
 
-    function mintToken(string memory tokenURI) public returns(uint) {
+    // We want to reduce our gas and that's why we using memory
+    function mintToken(string memory tokenURI) public returns (uint256) {
         _tokenIds.increment();
         uint256 newItemId = _tokenIds.current();
         _mint(msg.sender, newItemId);
@@ -31,7 +32,7 @@ contract NFT is ERC721URIStorage {
         _setTokenURI(newItemId, tokenURI);
         // give the marketplace the approval to transact between users
         setApprovalForAll(contractAddress, true);
-        // mint the token and set it for sale - return the id to do so 
+        // mint the token and set it for sale - return the id
         return newItemId;
     }
 }
